TLK-2091 check tool parameters generated by model - review fixes
