+++ src/backend/tools/base.py
@@ -1,7 +1,7 @@
 import datetime
 from abc import ABC, abstractmethod
 from enum import StrEnum
-from typing import Any, Dict, List, get_args, get_origin
+from typing import Any, Dict, List
 
 from fastapi import Request
 from pydantic import BaseModel
@@ -12,18 +12,22 @@ from backend.database_models.database import DBSessionDep
 from backend.database_models.tool_auth import ToolAuth
 from backend.schemas.tool import ToolDefinition
 from backend.services.logger.utils import LoggerFactory
+from backend.tools.utils.tools_checkers import check_tool_parameters
 
 logger = LoggerFactory().get_logger()
+
 
 class ToolErrorCode(StrEnum):
     HTTP_ERROR = "http_error"
     AUTH = "auth"
     OTHER = "other"
 
+
 class ToolAuthException(Exception):
     def __init__(self, message, tool_id: str):
         self.message = message
         self.tool_id = tool_id
+
 
 class ToolError(BaseModel, extra="allow"):
     type: ToolErrorCode = ToolErrorCode.OTHER
@@ -31,88 +35,22 @@ class ToolError(BaseModel, extra="allow"):
     text: str
     details: str = ""
 
-def check_type(param_value, type_description: str) -> bool:
-    try:
-        # Convert the type string into a type object
-        expected_type = eval(type_description)
-        return _check_type_recursive(param_value, expected_type)
-    except Exception as e:
-        print(f"Error during type checking: {e}")
-        return False
 
-def _check_type_recursive(value, expected_type) -> bool:
-    origin = get_origin(expected_type)
-
-    if origin is None:  # Base types (int, str, ...)
-        return isinstance(value, expected_type)
-
-    if origin is list:  # Check if the value is a list
-        if not isinstance(value, list):
-            return False
-        element_type = get_args(expected_type)[0]
-        return all(_check_type_recursive(item, element_type) for item in value)
-
-    if origin is tuple:  # Tuples
-        # trying to help to model with tuple type by converting lists to tuples, Cohere model passed tuples as list
-        converted_value = tuple(value) if isinstance(value, list) else value
-        if not isinstance(converted_value, tuple) or len(converted_value) != len(get_args(expected_type)):
-            return False
-        return all(
-            _check_type_recursive(item, arg_type)
-            for item, arg_type in zip(value, get_args(expected_type))
-        )
-
-    if origin is dict: # Dictionaries
-        if not isinstance(value, dict):
-            return False
-        key_type, value_type = get_args(expected_type)
-        return all(
-            _check_type_recursive(k, key_type) and _check_type_recursive(v, value_type)
-            for k, v in value.items()
-        )
-
-    # NOTE: Maybe we need to handle more types in the future, depends on the use in tools and models
-    return False
-
-def check_tool_parameters(tool_definition: ToolDefinition) -> None:
-    def decorator(func):
-        def wrapper(self, *args, **kwargs):
-            parameter_definitions = tool_definition(self).parameter_definitions
-            passed_method_params = kwargs.get("parameters", {}) or args[0]
-            # Validate parameters
-            for param, rules in parameter_definitions.items():
-                is_required = rules.get("required", False)
-                if param not in passed_method_params:
-                    if is_required:
-                        raise ValueError(f"Model didn't pass required parameter: {param}")
-                else:
-                    value = passed_method_params[param]
-                    if not value and is_required:
-                        raise ValueError(f"Model passed empty value for required parameter: {param}")
-                    if not check_type(value, rules["type"]):
-                        raise TypeError(
-                            f"Model passed invalid parameter. Parameter '{param}' must be of type {rules['type']}, but got {type(value).__name__}"
-                        )
-
-            return func(self, *args, **kwargs)
-
-        return wrapper
-
-    return decorator
-
-
-class ParametersCheckingMeta(type):
-    def __new__(cls, name, bases, dct):
-        for attr_name, attr_value in dct.items():
+class ParametersValidationMeta(type):
+    """
+    Metaclass to decorate all tools `call` methods with the parameter checker.
+    """
+    def __new__(cls, name, bases, class_dict):
+        for attr_name, attr_value in class_dict.items():
             if callable(attr_value) and attr_name == "call":
                 # Decorate methods with the parameter checker
-                dct[attr_name] = check_tool_parameters(
+                class_dict[attr_name] = check_tool_parameters(
                     lambda self: self.__class__.get_tool_definition()
                 )(attr_value)
-        return super().__new__(cls, name, bases, dct)
+        return super().__new__(cls, name, bases, class_dict)
 
 
-class BaseTool(metaclass=ParametersCheckingMeta):
+class BaseTool(metaclass=ParametersValidationMeta):
     """
     Abstract base class for all Tools.
 
@@ -130,11 +68,13 @@ class BaseTool(metaclass=ParametersCheckingMeta):
 
     @classmethod
     @abstractmethod
-    def is_available(cls) -> bool: ...
+    def is_available(cls) -> bool:
+        ...
 
     @classmethod
     @abstractmethod
-    def get_tool_definition(cls) -> ToolDefinition: ...
+    def get_tool_definition(cls) -> ToolDefinition:
+        ...
 
     @classmethod
     def generate_error_message(cls) -> str | None:
@@ -144,7 +84,8 @@ class BaseTool(metaclass=ParametersCheckingMeta):
         return f"{cls.__name__} is not available. Please make sure all required config variables are set."
 
     @classmethod
-    def _handle_tool_specific_errors(cls, error: Exception, **kwargs: Any) -> None: ...
+    def _handle_tool_specific_errors(cls, error: Exception, **kwargs: Any) -> None:
+        ...
 
     @classmethod
     def get_tool_error(cls, err: ToolError):
@@ -158,8 +99,9 @@ class BaseTool(metaclass=ParametersCheckingMeta):
 
     @abstractmethod
     async def call(
-        self, parameters: dict, ctx: Any, **kwargs: Any
-    ) -> List[Dict[str, Any]]: ...
+            self, parameters: dict, ctx: Any, **kwargs: Any
+    ) -> List[Dict[str, Any]]:
+        ...
 
 
 class BaseToolAuthentication(ABC):
@@ -176,11 +118,11 @@ class BaseToolAuthentication(ABC):
 
     def _post_init_check(self):
         if any(
-            [
-                self.BACKEND_HOST is None,
-                self.FRONTEND_HOST is None,
-                self.AUTH_SECRET_KEY is None,
-            ]
+                [
+                    self.BACKEND_HOST is None,
+                    self.FRONTEND_HOST is None,
+                    self.AUTH_SECRET_KEY is None,
+                ]
         ):
             raise ValueError(
                 "Tool Authentication requires auth.backend_hostname, auth.frontend_hostname in configuration.yaml, "
@@ -188,7 +130,8 @@ class BaseToolAuthentication(ABC):
             )
 
     @abstractmethod
-    def get_auth_url(self, user_id: str) -> str: ...
+    def get_auth_url(self, user_id: str) -> str:
+        ...
 
     def is_auth_required(self, session: DBSessionDep, user_id: str) -> bool:
         auth = tool_auth_crud.get_tool_auth(session, self.TOOL_ID, user_id)
@@ -221,13 +164,15 @@ class BaseToolAuthentication(ABC):
 
     @abstractmethod
     def try_refresh_token(
-        self, session: DBSessionDep, user_id: str, tool_auth: ToolAuth
-    ) -> bool: ...
+            self, session: DBSessionDep, user_id: str, tool_auth: ToolAuth
+    ) -> bool:
+        ...
 
     @abstractmethod
     def retrieve_auth_token(
-        self, request: Request, session: DBSessionDep, user_id: str
-    ) -> str: ...
+            self, request: Request, session: DBSessionDep, user_id: str
+    ) -> str:
+        ...
 
     def get_token(self, session: DBSessionDep, user_id: str) -> str:
         tool_auth = tool_auth_crud.get_tool_auth(session, self.TOOL_ID, user_id)
@@ -242,4 +187,3 @@ class BaseToolAuthentication(ABC):
                 event=f"BaseToolAuthentication: Error while deleting Tool Auth: {str(e)}"
             )
             raise
-
